Another Dumpy Commit
As the project is getting off the ground there will be a couple more
"dumpy" commits where large numbers of files are added and changed...
like this one. Nothing in the main code is operable still, but the
SpikeEvent handler is mostly done and uses the arrayfire library.
neuron.cpp and .h which will replace plasticneuron.c and h. Also added a
mex file that jitters spike trains and significant progress has been
made on the asdf2avlets code for finding causal paths. Finally netparams
has been updated with the most recent set of parameters for the network.