Today's work and updating gitignore
Biggest is that  MANA_Nodes now have a reform function so that synapses can be removed/added and also the constructor is now complete. All in all the class may actually be completely finished at this point.

A minor bug in sector was fixed which would've iterated over the wrong numbers for the evt_currents.